feat: create simple execution plan for entity call
